Set softtabstop=0 when noexpandtab is set
softtabstop is useful when expandtab is set because it makes the
tab key a useful tool for indenting. When noexpandtab is set,
it causes the tab, >, and < keys to insert mixed spaces for
indentation which isn't a useful behavior. Setting to 0 disables
this and allows true tab characters to be used.

Fixes #1